pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "synthstats"
version = "0.1.0"
description = "GFlowNet-steered probabilistic program synthesis for safer AI"
readme = "README.md"
requires-python = ">=3.12,<3.14"
authors = [
    {name = "Younesse Kaddar"},
    {name = "Sam Staton"},
    {name = "Nikolay Malkin"},
    {name = "Zihuiwen Ye"},
    {name = "Jacek Karwowski"},
]
dependencies = [
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "pandas>=2.2.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "tqdm>=4.67.0",
    "hydra-core>=1.3.0",
    "hydra-submitit-launcher>=1.2.0",
    "omegaconf>=2.3.0",
    "torch>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "hypothesis>=6.100.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "ipykernel>=6.30.0",
]
ml = [
    "transformers>=4.36.0",
    "peft>=0.7.0",
    "accelerate>=0.25.0",
    "tiktoken>=0.5.0",
    "litellm>=1.20.0",
]
pymc = [
    "pymc>=5.10.0",
    "arviz>=0.17.0",
]
gym = [
    "gymnasium>=0.29.0",
    "ray>=2.9.0",
]
wandb = [
    "wandb>=0.16.0",
]
tinker = [
    "tinker>=0.7.0",
    "tinker-cookbook>=0.1.0",
]
swe = [
    "docker>=7.0.0",
]
skyrl = [
    "skyrl-train; platform_system == 'Linux' and python_version >= '3.12' and python_version < '3.13'",
    "vllm>=0.6.0; platform_system == 'Linux' and python_version >= '3.12' and python_version < '3.13'",
]
training = [
    "bitsandbytes>=0.42.0; platform_system == 'Linux'",
    "peft>=0.7.0",
    "accelerate>=0.25.0",
]
large = [
    "megatron-core",
    "deepspeed",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocs-mermaid2-plugin>=1.1.0",
    "mkdocstrings[python]>=0.24.0",
    "mkdocs-git-revision-date-localized-plugin>=1.2.0",
    "mkdocs-minify-plugin>=0.8.0",
    "mkdocs-jupyter>=0.24.0",
    "ruff>=0.9.0",
]

[tool.uv.sources]
skyrl-train = { git = "https://github.com/NovaSky-AI/SkyRL.git", subdirectory = "skyrl-train", rev = "3fcbc05292a40ea0fe70ad81733f09eed49e5d58" }

# flash-attn requires torch at build time but doesn't declare it
[tool.uv.extra-build-dependencies]
flash-attn = ["torch"]

[project.scripts]
synthstats = "synthstats.cli:main"
synthstats-train = "synthstats.cli.train:main"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
include = ["synthstats*"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]

[tool.mypy]
python_version = "3.13"
ignore_missing_imports = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --tb=short"
markers = [
    "slow: marks tests as slow",
    "canary: opt-in integration tests (RUN_CANARY=1)",
    "integration: requires external services",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
